Improve list logic
Because reveals consistently appeared inside
of lists, a solution was to convert the contents
of lists into paragraphs, allow those paragraphs
to be broken into multiple paragraphs, and then
allow the markout scripts to operate on them.

Unfortunately, this left the HTML cluttered with
unecessary p-tags.  This update improves on that
solution by unwrapping the p-tags when only
a single child exists.

Solves: Issue #15