Updates all files to comply with basic jshint rules
The rules in the .jshintrc are yanked off of Stackoverflow without
much thought.  I did a first pass crossing T's dotting I's.  If
the changes required were minor enough, I refactored it on the spot.
There will be more changes required, especially as new rules are
added to the rc file.  In either case, this should have been done
a long time ago...