Significant refactor to mammothize module
Mammothize is the main function that converts the
docx into html.  It uses mammoth.js which is a little
difficult to manage due to it's use of promises among
other things.  The original solution was hacked together
and in dire need of repair.  This commit removes some of
the dark magic and makes it more clear what is happening.

There is probably some more work to come, but this is a
very promising start.