Simplifies the file writer
In some ways this is more complex because the preview page
generation is moved inside of the file writer at a point
where the html has already been templatized.  On the other
hand, it's simpler because there's no need to worry about
pulling the preview page out of the html array.

A further step might be to templatize inside the file writer.